Revise QLearning to only pick valid actions
If an MDP has states with differing number of actions,
the only way to represent the transition P matrix is
(A, S, S').  The problem is some rows for some actions
will be all 0's.  The before code may inadvertently pick
state transitions that had zero probability from occuring.
This code fixes it by requiring that the state transitions for
a valid action had to sum to 1.